requirements.txt
requests
fastapi
uvicorn[standard]
SQLAlchemy
python-decouple==3.8
gunicorn
psycopg2-binary
openai
PyJWT
bcrypt==4.0.1
passlib==1.7.4
pydantic
email-validator
redis==5.2.1
jinja2==3.1.3